requirements.txt
flask
langchain
langchain-google-genai
langchain-community
pymupdf
python-docx
beautifulsoup4
requests
numpy
faiss-cpu